+++ source/phantasy_engine/util/IOUtil.cpp
@@ -8,12 +8,12 @@ using sfz::DynString;
 
 DynString convertToOSPath(const char* path) noexcept
 {
-	size_t pathLen = std::strlen(path);
+	uint32_t pathLen = uint32_t(std::strlen(path));
 	DynString outPath;
 	outPath.setCapacity(uint32_t(pathLen + 1)); // +1 for null terminator
 
 	auto& internalArray = outPath.internalDynArray();
-	for (size_t i = 0; i < pathLen; i++) {
+	for (uint32_t i = 0; i < pathLen; i++) {
 		char c = path[i];
 #ifdef _WIN32
 		if (c == '/') {
